+++ typescript/web/src/components/dataset-class-list/__tests__/class-list.tsx
@@ -67,7 +67,9 @@ describe("Dataset class list tests", () => {
   it("Renders if the dataset has no classes", async () => {
     const datasetId = "myDatasetId";
     await createDataset("myDataset", datasetId);
-    render(<ClassesList datasetSlug="mydataset" />, { wrapper });
+    render(<ClassesList datasetSlug="mydataset" workspaceSlug="local" />, {
+      wrapper,
+    });
     expect(screen.getByText("0 Classes")).toBeDefined();
   });
 
@@ -89,7 +91,9 @@ describe("Dataset class list tests", () => {
       name: "MyThirdClass",
       color: "red",
     });
-    render(<ClassesList datasetSlug="mydataset" />, { wrapper });
+    render(<ClassesList datasetSlug="mydataset" workspaceSlug="local" />, {
+      wrapper,
+    });
 
     await waitFor(() => {
       expect(screen.getByText("3 Classes")).toBeDefined();
@@ -107,7 +111,9 @@ describe("Dataset class list tests", () => {
       name: "MyFirstClass",
       color: "blue",
     });
-    render(<ClassesList datasetSlug="mydataset" />, { wrapper });
+    render(<ClassesList datasetSlug="mydataset" workspaceSlug="local" />, {
+      wrapper,
+    });
 
     await waitFor(() =>
       fireEvent.click(screen.getByLabelText(/Delete class/i))