+++ typescript/web/src/components/labelling-tool/openlayers-map/labels.tsx
@@ -10,7 +10,7 @@ import { Fill, Stroke, Style } from "ol/style";
 
 import CircleStyle from "ol/style/Circle";
 import { Feature } from "ol";
-import { Label } from "@labelflow/graphql-types";
+import { Label, LabelType } from "@labelflow/graphql-types";
 
 import { useLabellingStore } from "../../../connectors/labelling-state";
 
@@ -20,7 +20,10 @@ const getImageLabelsQuery = gql`
   query getImageLabels($imageId: ID!) {
     image(where: { id: $imageId }) {
       id
+      width
+      height
       labels {
+        type
         id
         x
         y
@@ -28,6 +31,7 @@ const getImageLabelsQuery = gql`
         height
         labelClass {
           id
+          name
           color
         }
         geometry {
@@ -45,61 +49,65 @@ export const Labels = ({
   sourceVectorLabelsRef?: MutableRefObject<OlSourceVector<Geometry> | null>;
 }) => {
   const { imageId } = useRouter()?.query;
-  const { data } = useQuery(getImageLabelsQuery, {
+  const { data, previousData } = useQuery(getImageLabelsQuery, {
     skip: !imageId,
     variables: { imageId: imageId as string },
   });
   const selectedLabelId = useLabellingStore((state) => state.selectedLabelId);
-  const labels = data?.image?.labels ?? [];
+  const labels = data?.image?.labels ?? previousData?.image?.labels ?? [];
 
   return (
     <>
       <olLayerVector>
         <olSourceVector ref={sourceVectorLabelsRef}>
-          {labels.map(({ id, labelClass, geometry }: Label) => {
-            const isSelected = id === selectedLabelId;
-            const labelClassColor = labelClass?.color ?? noneClassColor;
-            const labelStyle = new Style({
-              fill: new Fill({
-                color: `${labelClassColor}${isSelected ? "40" : "10"}`,
-              }),
-              stroke: new Stroke({
-                color: labelClassColor,
-                width: isSelected ? 4 : 2,
-              }),
-              zIndex: isSelected ? 2 : 1,
-            });
-            const verticesStyle = isSelected
-              ? new Style({
-                  image: new CircleStyle({
-                    radius: 5,
-                    fill: new Fill({
-                      color: labelClassColor,
+          {labels
+            .filter(({ type }: Label) =>
+              [LabelType.Box, LabelType.Polygon].includes(type)
+            )
+            .map(({ id, labelClass, geometry }: Label) => {
+              const isSelected = id === selectedLabelId;
+              const labelClassColor = labelClass?.color ?? noneClassColor;
+              const labelStyle = new Style({
+                fill: new Fill({
+                  color: `${labelClassColor}${isSelected ? "40" : "10"}`,
+                }),
+                stroke: new Stroke({
+                  color: labelClassColor,
+                  width: isSelected ? 4 : 2,
+                }),
+                zIndex: isSelected ? 2 : 1,
+              });
+              const verticesStyle = isSelected
+                ? new Style({
+                    image: new CircleStyle({
+                      radius: 5,
+                      fill: new Fill({
+                        color: labelClassColor,
+                      }),
                     }),
-                  }),
-                  geometry: (feature) => {
-                    const coordinates = (feature as Feature<Polygon>)
-                      .getGeometry()
-                      .getCoordinates()[0];
-                    return new MultiPoint(coordinates);
-                  },
-                  zIndex: isSelected ? 2 : 1,
-                })
-              : null;
-            const style = isSelected
-              ? [labelStyle, verticesStyle]
-              : [labelStyle];
+                    geometry: (feature) => {
+                      const coordinates = (feature as Feature<Polygon>)
+                        .getGeometry()
+                        .getCoordinates()[0];
+                      return new MultiPoint(coordinates);
+                    },
+                    zIndex: isSelected ? 2 : 1,
+                  })
+                : null;
+              const style = isSelected
+                ? [labelStyle, verticesStyle]
+                : [labelStyle];
 
-            return (
-              <olFeature
-                key={id}
-                id={id}
-                properties={{ isSelected }}
-                geometry={new GeoJSON().readGeometry(geometry)}
-                style={style}
-              />
-            );
-          })}
+              return (
+                <olFeature
+                  key={id}
+                  id={id}
+                  properties={{ isSelected }}
+                  geometry={new GeoJSON().readGeometry(geometry)}
+                  style={style}
+                />
+              );
+            })}
         </olSourceVector>
       </olLayerVector>
     </>