+++ typescript/web/src/components/labelling-tool/options-tool-bar/edit-label-class-menu.tsx
@@ -1,20 +1,33 @@
-import { useMemo, useState, useEffect } from "react";
+import { useState, useEffect, useCallback } from "react";
 import { gql, useQuery, useApolloClient } from "@apollo/client";
+import { Label, LabelType } from "@labelflow/graphql-types";
+import GeoJSON, { GeoJSONPolygon } from "ol/format/GeoJSON";
+import { Polygon } from "ol/geom";
 
 import { useHotkeys } from "react-hotkeys-hook";
 import { useRouter } from "next/router";
 import { ClassSelectionMenu, LabelClassItem } from "./class-selection-menu";
+import { ClassAdditionMenu } from "./class-addition-menu";
 import { Tools, useLabellingStore } from "../../../connectors/labelling-state";
 import { useUndoStore } from "../../../connectors/undo-store";
-import { createNewLabelClassAndUpdateLabelCurry } from "../../../connectors/undo-store/effects/create-label-class-and-update-label";
+import { createCreateLabelClassAndUpdateLabelEffect } from "../../../connectors/undo-store/effects/create-label-class-and-update-label";
 import { createUpdateLabelClassOfLabelEffect } from "../../../connectors/undo-store/effects/update-label-class-of-label";
-import { createNewLabelClassCurry } from "../../../connectors/undo-store/effects/create-label-class";
+import { createCreateLabelClassEffect } from "../../../connectors/undo-store/effects/create-label-class";
+import {
+  getNextClassColor,
+  hexColorSequence,
+} from "../../../utils/class-color-generator";
 import { createUpdateLabelClassEffect } from "../../../connectors/undo-store/effects/update-label-class";
 import { keymap } from "../../../keymap";
 
+import { createCreateLabelEffect } from "../../../connectors/undo-store/effects/create-label";
+
+import { createDeleteLabelEffect } from "../../../connectors/undo-store/effects/delete-label";
+import { createCreateLabelClassAndCreateLabelEffect } from "../../../connectors/undo-store/effects/create-label-class-and-create-label";
+
 const getLabelClassesOfDatasetQuery = gql`
-  query getLabelClassesOfDataset($slug: String!) {
-    dataset(where: { slugs: { slug: $slug, workspaceSlug: "local" } }) {
+  query getLabelClassesOfDataset($slug: String!, $workspaceSlug: String!) {
+    dataset(where: { slugs: { slug: $slug, workspaceSlug: $workspaceSlug } }) {
       id
       labelClasses {
         id
@@ -35,10 +48,11 @@ const labelClassQuery = gql`
   }
 `;
 
-const labelQuery = gql`
+const getLabelQuery = gql`
   query getLabel($id: ID!) {
     label(where: { id: $id }) {
       id
+      type
       labelClass {
         id
         name
@@ -48,13 +62,45 @@ const labelQuery = gql`
   }
 `;
 
+const getImageLabelsQuery = gql`
+  query getImageLabels($imageId: ID!) {
+    image(where: { id: $imageId }) {
+      id
+      width
+      height
+      labels {
+        type
+        id
+        x
+        y
+        width
+        height
+        labelClass {
+          id
+          name
+          color
+        }
+        geometry {
+          type
+          coordinates
+        }
+      }
+    }
+  }
+`;
+
 export const EditLabelClassMenu = () => {
   const router = useRouter();
   const datasetSlug = router?.query.datasetSlug as string;
+
+  const workspaceSlug = router?.query.workspaceSlug as string;
+  const imageId = router?.query.imageId as string;
+
   const client = useApolloClient();
   const [isOpen, setIsOpen] = useState(false);
+
   const { data } = useQuery(getLabelClassesOfDatasetQuery, {
-    variables: { slug: datasetSlug },
+    variables: { slug: datasetSlug, workspaceSlug },
   });
   const datasetId = data?.dataset.id;
   const { perform } = useUndoStore();
@@ -64,7 +110,7 @@ export const EditLabelClassMenu = () => {
   );
   const selectedTool = useLabellingStore((state) => state.selectedTool);
   const selectedLabelId = useLabellingStore((state) => state.selectedLabelId);
-  const { data: selectedLabelData } = useQuery(labelQuery, {
+  const { data: selectedLabelData } = useQuery(getLabelQuery, {
     variables: { id: selectedLabelId },
     skip: selectedLabelId == null,
   });
@@ -75,6 +121,9 @@ export const EditLabelClassMenu = () => {
   const setSelectedLabelClassId = useLabellingStore(
     (state) => state.setSelectedLabelClassId
   );
+  const setSelectedLabelId = useLabellingStore(
+    (state) => state.setSelectedLabelId
+  );
 
   useEffect(() => {
     setSelectedLabelClassId(null);
@@ -84,53 +133,200 @@ export const EditLabelClassMenu = () => {
     variables: { id: selectedLabelClassId },
     skip: selectedLabelClassId == null,
   });
-  const isInDrawingMode = [Tools.BOX, Tools.POLYGON].includes(selectedTool);
+  const isInDrawingMode = [
+    Tools.CLASSIFICATION,
+    Tools.BOX,
+    Tools.POLYGON,
+  ].includes(selectedTool);
   const selectedLabelClass = isInDrawingMode
     ? dataLabelClass?.labelClass
     : selectedLabelData?.label?.labelClass;
-  const createNewClass = useMemo(
-    () =>
-      createNewLabelClassCurry({
-        labelClasses,
-        datasetId,
-        datasetSlug,
-        perform,
-        client,
-      }),
-    [labelClasses, datasetId, selectedTool]
+
+  const handleCreateNewClass = useCallback(
+    async (name) => {
+      const newClassColor =
+        labelClasses.length < 1
+          ? hexColorSequence[0]
+          : getNextClassColor(labelClasses[labelClasses.length - 1].color);
+      if (!isInDrawingMode) {
+        // Update class of an existing label with a new class
+        await perform(
+          createCreateLabelClassAndUpdateLabelEffect(
+            {
+              name,
+              color: newClassColor,
+              selectedLabelId,
+              datasetId,
+              datasetSlug,
+              workspaceSlug,
+            },
+            { client }
+          )
+        );
+        return;
+      }
+
+      if (selectedTool === Tools.CLASSIFICATION && imageId) {
+        // Create a new classification label of a new class
+        const { data: imageLabelsData } = await client.query({
+          query: getImageLabelsQuery,
+          variables: { imageId },
+        });
+        const geometry = new GeoJSON().writeGeometryObject(
+          new Polygon([
+            [
+              [0, 0],
+              [0, imageLabelsData.image.height],
+              [imageLabelsData.image.width, imageLabelsData.image.height],
+              [imageLabelsData.image.width, 0],
+              [0, 0],
+            ],
+          ])
+        ) as GeoJSONPolygon;
+
+        await perform(
+          createCreateLabelClassAndCreateLabelEffect(
+            {
+              name,
+              color: newClassColor,
+              datasetId,
+              datasetSlug,
+              imageId,
+              previouslySelectedLabelClassId: selectedLabelClassId,
+              geometry,
+              labelType: LabelType.Classification,
+              workspaceSlug,
+            },
+            {
+              setSelectedLabelId,
+              client,
+            }
+          )
+        );
+        return;
+      }
+
+      // Change currently select class in the menu to an existing class, don't apply the class to any label
+      await perform(
+        createCreateLabelClassEffect(
+          {
+            name,
+            color: newClassColor,
+            selectedLabelClassIdPrevious: selectedLabelClassId,
+            datasetId,
+            datasetSlug,
+            workspaceSlug,
+          },
+          { client }
+        )
+      );
+    },
+    [labelClasses, selectedLabelId, datasetId, selectedTool, workspaceSlug]
   );
-  const createNewClassAndUpdateLabel = useMemo(
-    () =>
-      createNewLabelClassAndUpdateLabelCurry({
-        labelClasses,
-        datasetId,
-        datasetSlug,
-        perform,
-        client,
-      }),
-    [labelClasses, datasetId]
-  );
-  const onSelectedClassChange = useMemo(
-    () =>
-      isInDrawingMode
-        ? (item: LabelClassItem | null) =>
-            perform(
-              createUpdateLabelClassEffect({
+
+  const handleSelectedClassChange = useCallback(
+    async (item: LabelClassItem | null) => {
+      if (!isInDrawingMode) {
+        if (selectedLabelId != null) {
+          if (selectedLabelData?.label?.type === LabelType.Classification) {
+            // Change the class of an existing classification label to an existing class
+            const { data: imageLabelsData } = await client.query({
+              query: getImageLabelsQuery,
+              variables: { imageId },
+            });
+
+            const classificationsOfThisClass =
+              imageLabelsData.image.labels.filter(
+                (label: Label) =>
+                  label.labelClass?.id === item?.id &&
+                  label.type === LabelType.Classification
+              );
+            if (classificationsOfThisClass.length > 0) {
+              // If there is already a classification of the same class, delete the current one (to merge them)
+              await perform(
+                createDeleteLabelEffect(
+                  { id: selectedLabelId },
+                  { setSelectedLabelId, client }
+                )
+              );
+              return;
+            }
+          }
+
+          // Change the class of an existing label to an existing class
+          await perform(
+            createUpdateLabelClassOfLabelEffect(
+              {
                 selectedLabelClassId: item?.id ?? null,
-                selectedLabelClassIdPrevious: selectedLabelClassId,
-              })
+                selectedLabelId,
+              },
+              { client }
             )
-        : (item: LabelClassItem | null) =>
-            perform(
-              createUpdateLabelClassOfLabelEffect(
-                {
-                  selectedLabelClassId: item?.id ?? null,
-                  selectedLabelId,
-                },
-                { client }
-              )
-            ),
-    [selectedLabelId, selectedLabelClassId, selectedTool]
+          );
+          return;
+        }
+        return;
+      }
+      if (selectedTool === Tools.CLASSIFICATION && imageId) {
+        // Add a classification label of an existing class
+        const { data: imageLabelsData } = await client.query({
+          query: getImageLabelsQuery,
+          variables: { imageId },
+        });
+
+        const classificationsOfThisClass = imageLabelsData.image.labels.filter(
+          (label: Label) =>
+            label.labelClass?.id === item?.id &&
+            label.type === LabelType.Classification
+        );
+        if (classificationsOfThisClass.length > 0) {
+          // If there is already a classification of the same class, delete it (to toggle the classification label on/off)
+          await perform(
+            createDeleteLabelEffect(
+              { id: classificationsOfThisClass[0].id },
+              { setSelectedLabelId, client }
+            )
+          );
+          return;
+        }
+
+        const geometry = new GeoJSON().writeGeometryObject(
+          new Polygon([
+            [
+              [0, 0],
+              [0, imageLabelsData.image.height],
+              [imageLabelsData.image.width, imageLabelsData.image.height],
+              [imageLabelsData.image.width, 0],
+              [0, 0],
+            ],
+          ])
+        ) as GeoJSONPolygon;
+
+        await perform(
+          createCreateLabelEffect(
+            {
+              imageId,
+              selectedLabelClassId: item?.id ?? null,
+              geometry,
+              labelType: LabelType.Classification,
+            },
+            {
+              setSelectedLabelId,
+              client,
+            }
+          )
+        );
+        return;
+      }
+      // Change currently select class in the menu to an existing class, don't apply the class to any label
+      await perform(
+        createUpdateLabelClassEffect({
+          selectedLabelClassId: item?.id ?? null,
+          selectedLabelClassIdPrevious: selectedLabelClassId,
+        })
+      );
+    },
+    [selectedLabelId, selectedLabelClassId, selectedTool, imageId]
   );
 
   const displayClassSelectionMenu =
@@ -148,31 +344,39 @@ export const EditLabelClassMenu = () => {
         const digit = Number(keyboardEvent.code[5]);
         const indexOfLabelClass = (digit + 9) % 10;
         if (indexOfLabelClass < labelClasses.length) {
-          onSelectedClassChange(labelClasses[indexOfLabelClass]);
+          handleSelectedClassChange(labelClasses[indexOfLabelClass]);
           setIsOpen(false);
         }
       }
     },
     {},
-    [labelClasses, onSelectedClassChange, isContextMenuOpen, setIsOpen]
+    [labelClasses, handleSelectedClassChange, isContextMenuOpen, setIsOpen]
   );
 
   return (
     <>
-      {displayClassSelectionMenu && (
+      {displayClassSelectionMenu && selectedTool !== Tools.CLASSIFICATION && (
         <ClassSelectionMenu
           isOpen={isOpen}
           setIsOpen={setIsOpen}
           selectedLabelClass={selectedLabelClass}
           labelClasses={labelClasses}
-          createNewClass={async (name) => {
-            return isInDrawingMode
-              ? await createNewClass(name, selectedLabelClassId)
-              : await createNewClassAndUpdateLabel(name, selectedLabelId);
-          }}
-          onSelectedClassChange={(item) => {
-            onSelectedClassChange(item);
-          }}
+          createNewClass={handleCreateNewClass}
+          onSelectedClassChange={handleSelectedClassChange}
+          includeNoneClass={
+            selectedLabelData?.label?.type !== LabelType.Classification
+          }
+          isContextMenuOpen={isContextMenuOpen}
+        />
+      )}
+      {displayClassSelectionMenu && selectedTool === Tools.CLASSIFICATION && (
+        <ClassAdditionMenu
+          isOpen={isOpen}
+          setIsOpen={setIsOpen}
+          selectedLabelClass={selectedLabelClass}
+          labelClasses={labelClasses}
+          createNewClass={handleCreateNewClass}
+          onSelectedClassChange={handleSelectedClassChange}
           isContextMenuOpen={isContextMenuOpen}
         />
       )}