+++ typescript/web/src/connectors/undo-store/effects/__tests__/create-label-class-and-update-label.tsx
@@ -52,6 +52,7 @@ beforeEach(async () => {
         color: "0xaa45f8",
         datasetId: "a dataset id",
         datasetSlug: "a-dataset-slug",
+        workspaceSlug: "local",
         selectedLabelId: "my label id",
       },
       { client }