+++ typescript/web/src/connectors/undo-store/effects/create-label-class-and-update-label.ts
@@ -1,16 +1,12 @@
 import { gql, ApolloClient } from "@apollo/client";
 
-import { LabelClass } from "@labelflow/graphql-types";
 import { useLabellingStore } from "../../labelling-state";
-import {
-  getNextClassColor,
-  hexColorSequence,
-} from "../../../utils/class-color-generator";
+
 import { Effect } from "..";
-import { getDatasetsQuery } from "../../../pages/local/datasets";
+import { getDatasetsQuery } from "../../../pages/[workspaceSlug]/datasets";
 import { datasetLabelClassesQuery } from "../../../components/dataset-class-list/class-item";
 
-const labelQuery = gql`
+const getLabelQuery = gql`
   query getLabel($id: ID!) {
     label(where: { id: $id }) {
       id
@@ -54,12 +50,14 @@ export const createCreateLabelClassAndUpdateLabelEffect = (
     color,
     datasetId,
     datasetSlug,
+    workspaceSlug,
     selectedLabelId,
   }: {
     name: string;
     color: string;
     datasetId: string;
     datasetSlug: string;
+    workspaceSlug: string;
     selectedLabelId: string | null;
   },
   {
@@ -79,7 +77,10 @@ export const createCreateLabelClassAndUpdateLabelEffect = (
       refetchQueries: [
         "getLabelClassesOfDataset",
         { query: getDatasetsQuery },
-        { query: datasetLabelClassesQuery, variables: { slug: datasetSlug } },
+        {
+          query: datasetLabelClassesQuery,
+          variables: { slug: datasetSlug, workspaceSlug },
+        },
       ],
     });
 
@@ -88,7 +89,7 @@ export const createCreateLabelClassAndUpdateLabelEffect = (
         label: { labelClass },
       },
     } = await client.query({
-      query: labelQuery,
+      query: getLabelQuery,
       variables: { id: selectedLabelId },
     });
 
@@ -132,7 +133,10 @@ export const createCreateLabelClassAndUpdateLabelEffect = (
       refetchQueries: [
         "getLabelClassesOfDataset",
         { query: getDatasetsQuery },
-        { query: datasetLabelClassesQuery, variables: { slug: datasetSlug } },
+        {
+          query: datasetLabelClassesQuery,
+          variables: { slug: datasetSlug, workspaceSlug },
+        },
       ],
     });
 
@@ -155,7 +159,10 @@ export const createCreateLabelClassAndUpdateLabelEffect = (
       refetchQueries: [
         "getLabelClassesOfDataset",
         { query: getDatasetsQuery },
-        { query: datasetLabelClassesQuery, variables: { slug: datasetSlug } },
+        {
+          query: datasetLabelClassesQuery,
+          variables: { slug: datasetSlug, workspaceSlug },
+        },
       ],
     });
 
@@ -175,33 +182,3 @@ export const createCreateLabelClassAndUpdateLabelEffect = (
     };
   },
 });
-
-export const createNewLabelClassAndUpdateLabelCurry =
-  ({
-    labelClasses,
-    datasetId,
-    datasetSlug,
-    perform,
-    onClose = () => {},
-    client,
-  }: {
-    labelClasses: LabelClass[];
-    datasetId: string;
-    datasetSlug: string;
-    perform: any;
-    onClose?: () => void;
-    client: ApolloClient<object>;
-  }) =>
-  async (name: string, selectedLabelId: string | null) => {
-    const newClassColor =
-      labelClasses.length < 1
-        ? hexColorSequence[0]
-        : getNextClassColor(labelClasses[labelClasses.length - 1].color);
-    perform(
-      createCreateLabelClassAndUpdateLabelEffect(
-        { name, color: newClassColor, selectedLabelId, datasetId, datasetSlug },
-        { client }
-      )
-    );
-    onClose();
-  };