+++ typescript/web/src/pages/[workspaceSlug]/datasets/[datasetSlug]/images/index.tsx
@@ -39,8 +39,8 @@ import { CookieBanner } from "../../../../../components/cookie-banner";
 const TrashIcon = chakra(HiTrash);
 
 export const datasetDataQuery = gql`
-  query getDatasetData($slug: String!) {
-    dataset(where: { slugs: { slug: $slug, workspaceSlug: "local" } }) {
+  query getDatasetData($slug: String!, $workspaceSlug: String!) {
+    dataset(where: { slugs: { slug: $slug, workspaceSlug: $workspaceSlug } }) {
       id
       name
       images {
@@ -63,6 +63,7 @@ const deleteImageQuery = gql`
 const ImagesPage = () => {
   const router = useRouter();
   const datasetSlug = router?.query?.datasetSlug as string;
+  const workspaceSlug = router?.query?.workspaceSlug as string;
 
   const {
     data: datasetResult,
@@ -73,8 +74,9 @@ const ImagesPage = () => {
   }>(datasetDataQuery, {
     variables: {
       slug: datasetSlug,
+      workspaceSlug,
     },
-    skip: !datasetSlug,
+    skip: !datasetSlug || !workspaceSlug,
   });
 
   const [deleteImage] = useMutation(deleteImageQuery);
@@ -109,15 +111,15 @@ const ImagesPage = () => {
       <CookieBanner />
       <Layout
         breadcrumbs={[
-          <NextLink href="/local/datasets">
+          <NextLink key={0} href={`/${workspaceSlug}/datasets`}>
             <BreadcrumbLink>Datasets</BreadcrumbLink>
           </NextLink>,
-          <NextLink href={`/local/datasets/${datasetSlug}`}>
+          <NextLink key={1} href={`/${workspaceSlug}/datasets/${datasetSlug}`}>
             <BreadcrumbLink>
               {datasetName ?? <Skeleton>Dataset Name</Skeleton>}
             </BreadcrumbLink>
           </NextLink>,
-          <Text>Images</Text>,
+          <Text key={2}>Images</Text>,
         ]}
         topBarRightContent={
           <>
@@ -126,7 +128,13 @@ const ImagesPage = () => {
             <ExportButton />
           </>
         }
-        tabBar={<DatasetTabBar currentTab="images" datasetSlug={datasetSlug} />}
+        tabBar={
+          <DatasetTabBar
+            currentTab="images"
+            datasetSlug={datasetSlug}
+            workspaceSlug={workspaceSlug}
+          />
+        }
       >
         {!datasetResult && (
           <Center h="full">
@@ -168,10 +176,12 @@ const ImagesPage = () => {
           >
             {datasetResult?.dataset?.images?.map(({ id, name, url }) => (
               <NextLink
-                href={`/local/datasets/${datasetSlug}/images/${id}`}
+                href={`/${workspaceSlug}/datasets/${datasetSlug}/images/${id}`}
                 key={id}
               >
-                <a href={`/local/datasets/${datasetSlug}/images/${id}`}>
+                <a
+                  href={`/${workspaceSlug}/datasets/${datasetSlug}/images/${id}`}
+                >
                   <VStack
                     maxW="486px"
                     p={4}